@@ -20,25 +20,25 @@
   "private": true,
   "dependencies": {
     "@abraham/reflection": "0.5.1",
-    "@angular/animations": "7.2.15",
-    "@angular/cdk": "8.0.0",
-    "@angular/common": "7.2.15",
-    "@angular/compiler": "7.2.15",
-    "@angular/core": "7.2.15",
-    "@angular/forms": "7.2.15",
+    "@angular/animations": "8.0.0",
+    "@angular/cdk": "8.0.1",
+    "@angular/common": "8.0.0",
+    "@angular/compiler": "8.0.0",
+    "@angular/core": "8.0.0",
+    "@angular/forms": "8.0.0",
     "@angular/http": "7.2.15",
-    "@angular/platform-browser": "7.2.15",
-    "@angular/platform-browser-dynamic": "7.2.15",
-    "@angular/router": "7.2.15",
+    "@angular/platform-browser": "8.0.0",
+    "@angular/platform-browser-dynamic": "8.0.0",
+    "@angular/router": "8.0.0",
     "@firebase/app": "0.4.4",
-    "@firebase/firestore": "1.3.3",
+    "@firebase/firestore": "1.3.5",
     "@webcomponents/webcomponentsjs": "2.2.10",
     "autotrack": "2.4.1",
-    "core-js": "2.6.9",
+    "core-js": "3.1.3",
     "dayjs": "1.8.14",
     "firebase-admin": "8.0.0",
     "github-repository": "0.3.0",
-    "material-components-web": "0.44.1",
+    "material-components-web": "2.3.0",
     "ng-lazyload-image": "5.1.2",
     "node-package": "1.3.0",
     "rxjs": "6.5.2",
@@ -46,14 +46,14 @@
     "zone.js": "0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.13.9",
-    "@angular/cli": "7.3.9",
-    "@angular/compiler-cli": "7.2.15",
-    "@angular/language-service": "7.2.15",
+    "@angular-devkit/build-angular": "0.800.2",
+    "@angular/cli": "^8.0.2",
+    "@angular/compiler-cli": "8.0.0",
+    "@angular/language-service": "8.0.0",
     "@types/jasmine": "3.3.13",
     "@types/jasminewd2": "2.0.6",
     "@types/material-components-web": "0.43.1",
-    "@types/node": "10.12.24",
+    "@types/node": "12.0.7",
     "@types/twit": "2.2.22",
     "codelyzer": "5.1.0",
     "firebase-tools": "6.11.0",
@@ -72,7 +72,7 @@
     "ts-node": "8.2.0",
     "tslint": "5.17.0",
     "twit": "2.2.11",
-    "typescript": "3.2.4",
+    "typescript": "^3.4.5",
     "workbox-build": "4.3.1",
     "workbox-sw": "4.3.1"
   }
