@@ -2,22 +2,18 @@ import { AfterViewInit, Component, ElementRef, ViewChild } from '@angular/core';
 import { MDCDialog } from '@material/dialog/index';
 import { MDCRipple } from '@material/ripple/index';
 
-type Dialog = MDCDialog & {
-  open: () => void;
-};
-
 @Component({
   selector: 'app-invite',
   templateUrl: './invite.component.html',
   styleUrls: ['./invite.component.scss']
 })
 export class InviteComponent implements AfterViewInit {
-  private dialog!: Dialog;
-  @ViewChild('fabEl') fabEl!: ElementRef;
-  @ViewChild('dialogEl') dialogEl!: ElementRef;
+  private dialog!: MDCDialog;
+  @ViewChild('fabEl', { static: true }) fabEl!: ElementRef;
+  @ViewChild('dialogEl', { static: true }) dialogEl!: ElementRef;
 
   ngAfterViewInit() {
-    this.dialog = new MDCDialog(this.dialogEl.nativeElement) as Dialog;
+    this.dialog = new MDCDialog(this.dialogEl.nativeElement);
     this.dialog.listen('MDCDialog:closing', () => this.fabEl.nativeElement.classList.remove('mdc-fab--exited'));
     this.initRipples();
   }
